added AuthUserTests tests renamed according to resources path
