--- src/test/java/EditUserTests.java
@@ -1,4 +1,0 @@
-public class EditUserTests {
-	//PATCH https://stellarburgers.nomoreparties.site/api/auth/user — эндпоинт обновления
-	//данных о пользователе
-}
